nyaysaathi/backend/requirements.txt
django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1
pymongo==4.7.3
python-dotenv==1.0.1
scikit-learn==1.5.0
numpy==1.26.4
gunicorn==22.0.0
whitenoise==6.7.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rapidfuzz==3.9.6
openai==1.54.4
langdetect==1.0.9
